fix(engine): 정보 전달 추가 리뷰 반영
