+++ apps/server/internal/engine/phase_engine_hooks_test.go
@@ -1,0 +1,146 @@
+package engine
+
+import (
+	"context"
+	"encoding/json"
+	"testing"
+)
+
+type hookRecordingModule struct {
+	stubCoreModule
+	entered []Phase
+	exited  []Phase
+}
+
+func (h *hookRecordingModule) OnPhaseEnter(_ context.Context, phase Phase) error {
+	h.entered = append(h.entered, phase)
+	return nil
+}
+
+func (h *hookRecordingModule) OnPhaseExit(_ context.Context, phase Phase) error {
+	h.exited = append(h.exited, phase)
+	return nil
+}
+
+func TestPhaseEngine_PhaseHooksAndOnEnterActions(t *testing.T) {
+	reactor := &stubFullModule{
+		stubCoreModule: stubCoreModule{name: "information_delivery"},
+		actions:        []PhaseAction{ActionDeliverInformation},
+	}
+	hook := &hookRecordingModule{stubCoreModule: stubCoreModule{name: "hook"}}
+	phases := []PhaseDefinition{
+		{
+			ID:      "intro",
+			Name:    "Intro",
+			OnEnter: json.RawMessage(`[{"type":"DELIVER_INFORMATION","params":{"deliveries":[{"id":"d1"}]}}]`),
+			OnExit:  json.RawMessage(`[{"type":"DELIVER_INFORMATION","params":{"deliveries":[{"id":"d2"}]}}]`),
+		},
+		{ID: "next", Name: "Next"},
+	}
+	pe, _ := newTestPhaseEngine(t, []Module{hook, reactor}, phases)
+	ctx := context.Background()
+	if err := pe.Start(ctx, nil); err != nil {
+		t.Fatalf("Start: %v", err)
+	}
+	defer pe.Stop(ctx)
+
+	if len(hook.entered) != 1 || hook.entered[0] != "intro" {
+		t.Fatalf("entered hooks = %#v", hook.entered)
+	}
+	if len(reactor.received) != 1 || reactor.received[0].Action != ActionDeliverInformation {
+		t.Fatalf("received actions = %#v", reactor.received)
+	}
+
+	if _, err := pe.AdvancePhase(ctx); err != nil {
+		t.Fatalf("AdvancePhase: %v", err)
+	}
+	if len(reactor.received) != 2 || reactor.received[1].Action != ActionDeliverInformation {
+		t.Fatalf("expected OnEnter + OnExit actions, got %#v", reactor.received)
+	}
+	if len(hook.exited) != 1 || hook.exited[0] != "intro" {
+		t.Fatalf("exited hooks = %#v", hook.exited)
+	}
+	if len(hook.entered) != 2 || hook.entered[1] != "next" {
+		t.Fatalf("entered hooks after advance = %#v", hook.entered)
+	}
+}
+
+func TestPhaseEngine_LegacyJSONLogicOnEnterIsNoop(t *testing.T) {
+	pe, _ := newTestPhaseEngine(t, nil, []PhaseDefinition{
+		{ID: "intro", Name: "Intro", OnEnter: json.RawMessage(`{"==":[1,1]}`)},
+	})
+	if err := pe.Start(context.Background(), nil); err != nil {
+		t.Fatalf("legacy JSONLogic-ish onEnter should remain no-op, got: %v", err)
+	}
+	defer pe.Stop(context.Background())
+}
+
+type panicHookModule struct {
+	stubCoreModule
+	panicOnEnter bool
+	panicOnExit  bool
+}
+
+func (p *panicHookModule) OnPhaseEnter(_ context.Context, _ Phase) error {
+	if p.panicOnEnter {
+		panic("enter hook boom")
+	}
+	return nil
+}
+
+func (p *panicHookModule) OnPhaseExit(_ context.Context, _ Phase) error {
+	if p.panicOnExit {
+		panic("exit hook boom")
+	}
+	return nil
+}
+
+func TestPhaseEngine_RequiredModuleDispatchUsesPanicIsolation(t *testing.T) {
+	pm := &panicModule{stubCoreModule: stubCoreModule{name: "information_delivery"}}
+	pe, audit := newTestPhaseEngine(t, []Module{pm}, testPhaseDefinitions)
+	ctx := context.Background()
+	if err := pe.Start(ctx, nil); err != nil {
+		t.Fatal(err)
+	}
+	defer pe.Stop(ctx)
+
+	err := pe.DispatchAction(ctx, PhaseActionPayload{Action: ActionDeliverInformation})
+	if err == nil {
+		t.Fatal("expected panic-isolated error")
+	}
+	if got := len(audit.eventsOfType("module.panic")); got != 1 {
+		t.Fatalf("module.panic audits = %d", got)
+	}
+}
+
+func TestPhaseEngine_PhaseHookPanicIsolation(t *testing.T) {
+	pm := &panicHookModule{stubCoreModule: stubCoreModule{name: "hook_panic"}, panicOnEnter: true}
+	pe, audit := newTestPhaseEngine(t, []Module{pm}, testPhaseDefinitions)
+	err := pe.Start(context.Background(), nil)
+	if err == nil {
+		t.Fatal("expected hook panic to become error")
+	}
+	if got := len(audit.eventsOfType("module.panic")); got != 1 {
+		t.Fatalf("module.panic audits = %d", got)
+	}
+}
+
+func TestPhaseEngine_SkipToPhaseRunsExitHooks(t *testing.T) {
+	hook := &hookRecordingModule{stubCoreModule: stubCoreModule{name: "hook"}}
+	pe, _ := newTestPhaseEngine(t, []Module{hook}, testPhaseDefinitions)
+	ctx := context.Background()
+	if err := pe.Start(ctx, nil); err != nil {
+		t.Fatal(err)
+	}
+	defer pe.Stop(ctx)
+
+	if err := pe.SkipToPhase(ctx, "vote"); err != nil {
+		t.Fatal(err)
+	}
+	if len(hook.exited) != 1 || hook.exited[0] != "intro" {
+		t.Fatalf("skip should exit old phase, exited=%#v", hook.exited)
+	}
+	if len(hook.entered) != 2 || hook.entered[1] != "vote" {
+		t.Fatalf("skip should enter target phase, entered=%#v", hook.entered)
+	}
+}
