+++ apps/server/internal/engine/phase_engine_test.go
@@ -487,307 +487,3 @@ func TestPhaseEngine_EventBusPhaseAdvance(t *testing.T) {
 		t.Fatal("expected phase:entered event for invest")
 	}
 }
-
-func TestPhaseEngine_NilAuditLogger(t *testing.T) {
-	// Passing nil audit logger should not panic — uses noopAuditLogger.
-	logger := &testLogger{t}
-	bus := NewEventBus(logger)
-	pe := NewPhaseEngine(uuid.New(), nil, bus, nil, logger, testPhaseDefinitions)
-
-	ctx := context.Background()
-	if err := pe.Start(ctx, nil); err != nil {
-		t.Fatal(err)
-	}
-	pe.Stop(ctx)
-}
-
-func TestPhaseEngine_ModuleConfigPassedToInit(t *testing.T) {
-	var receivedConfig json.RawMessage
-	mod := &configCapture{stubCoreModule: stubCoreModule{name: "cfg_mod"}, onInit: func(cfg json.RawMessage) {
-		receivedConfig = cfg
-	}}
-	pe, _ := newTestPhaseEngine(t, []Module{mod}, testPhaseDefinitions)
-
-	configs := map[string]json.RawMessage{
-		"cfg_mod": json.RawMessage(`{"key":"value"}`),
-	}
-	if err := pe.Start(context.Background(), configs); err != nil {
-		t.Fatal(err)
-	}
-	defer pe.Stop(context.Background())
-
-	if string(receivedConfig) != `{"key":"value"}` {
-		t.Fatalf("expected config passed through, got %s", receivedConfig)
-	}
-}
-
-// configCapture is a stub that captures the config passed to Init.
-type configCapture struct {
-	stubCoreModule
-	onInit func(json.RawMessage)
-}
-
-func (c *configCapture) Init(ctx context.Context, deps ModuleDeps, config json.RawMessage) error {
-	if c.onInit != nil {
-		c.onInit(config)
-	}
-	return c.stubCoreModule.Init(ctx, deps, config)
-}
-
-// Phase 20 PR-5: CurrentRound — monotonic counter starting at 1 on Start,
-// incrementing once per successful AdvancePhase, staying put after the final
-// phase exits.
-func TestPhaseEngine_CurrentRound_IncrementsWithAdvance(t *testing.T) {
-	pe, _ := newTestPhaseEngine(t, nil, testPhaseDefinitions)
-	ctx := context.Background()
-
-	if got := pe.CurrentRound(); got != 0 {
-		t.Fatalf("pre-Start round = %d, want 0", got)
-	}
-
-	if err := pe.Start(ctx, nil); err != nil {
-		t.Fatal(err)
-	}
-	defer pe.Stop(ctx)
-
-	if got := pe.CurrentRound(); got != 1 {
-		t.Fatalf("post-Start round = %d, want 1", got)
-	}
-
-	// Advance 1 → round 2
-	if _, err := pe.AdvancePhase(ctx); err != nil {
-		t.Fatalf("advance 1: %v", err)
-	}
-	if got := pe.CurrentRound(); got != 2 {
-		t.Fatalf("after advance 1, round = %d, want 2", got)
-	}
-
-	// Advance 2 → round 3
-	if _, err := pe.AdvancePhase(ctx); err != nil {
-		t.Fatalf("advance 2: %v", err)
-	}
-	if got := pe.CurrentRound(); got != 3 {
-		t.Fatalf("after advance 2, round = %d, want 3", got)
-	}
-
-	// Advance past the last phase — engine completes, round sticks.
-	if hasNext, _ := pe.AdvancePhase(ctx); hasNext {
-		t.Fatal("expected engine complete on 3rd advance")
-	}
-	if got := pe.CurrentRound(); got != 3 {
-		t.Fatalf("post-completion round = %d, want 3 (sticks)", got)
-	}
-}
-
-type hookRecordingModule struct {
-	stubCoreModule
-	entered []Phase
-	exited  []Phase
-}
-
-func (h *hookRecordingModule) OnPhaseEnter(_ context.Context, phase Phase) error {
-	h.entered = append(h.entered, phase)
-	return nil
-}
-
-func (h *hookRecordingModule) OnPhaseExit(_ context.Context, phase Phase) error {
-	h.exited = append(h.exited, phase)
-	return nil
-}
-
-func TestPhaseEngine_PhaseHooksAndOnEnterActions(t *testing.T) {
-	reactor := &stubFullModule{
-		stubCoreModule: stubCoreModule{name: "information_delivery"},
-		actions:        []PhaseAction{ActionDeliverInformation},
-	}
-	hook := &hookRecordingModule{stubCoreModule: stubCoreModule{name: "hook"}}
-	phases := []PhaseDefinition{
-		{
-			ID:      "intro",
-			Name:    "Intro",
-			OnEnter: json.RawMessage(`[{"type":"DELIVER_INFORMATION","params":{"deliveries":[{"id":"d1"}]}}]`),
-		},
-		{ID: "next", Name: "Next"},
-	}
-	pe, _ := newTestPhaseEngine(t, []Module{hook, reactor}, phases)
-	ctx := context.Background()
-	if err := pe.Start(ctx, nil); err != nil {
-		t.Fatalf("Start: %v", err)
-	}
-	defer pe.Stop(ctx)
-
-	if len(hook.entered) != 1 || hook.entered[0] != "intro" {
-		t.Fatalf("entered hooks = %#v", hook.entered)
-	}
-	if len(reactor.received) != 1 || reactor.received[0].Action != ActionDeliverInformation {
-		t.Fatalf("received actions = %#v", reactor.received)
-	}
-
-	if _, err := pe.AdvancePhase(ctx); err != nil {
-		t.Fatalf("AdvancePhase: %v", err)
-	}
-	if len(hook.exited) != 1 || hook.exited[0] != "intro" {
-		t.Fatalf("exited hooks = %#v", hook.exited)
-	}
-	if len(hook.entered) != 2 || hook.entered[1] != "next" {
-		t.Fatalf("entered hooks after advance = %#v", hook.entered)
-	}
-}
-
-func TestPhaseEngine_LegacyJSONLogicOnEnterIsNoop(t *testing.T) {
-	pe, _ := newTestPhaseEngine(t, nil, []PhaseDefinition{
-		{ID: "intro", Name: "Intro", OnEnter: json.RawMessage(`{"==":[1,1]}`)},
-	})
-	if err := pe.Start(context.Background(), nil); err != nil {
-		t.Fatalf("legacy JSONLogic-ish onEnter should remain no-op, got: %v", err)
-	}
-	defer pe.Stop(context.Background())
-}
-
-type failingEnterHookModule struct {
-	stubCoreModule
-	failPhase Phase
-}
-
-func (h *failingEnterHookModule) OnPhaseEnter(_ context.Context, phase Phase) error {
-	if phase == h.failPhase {
-		return fmt.Errorf("enter failed for %s", phase)
-	}
-	return nil
-}
-
-func (h *failingEnterHookModule) OnPhaseExit(_ context.Context, _ Phase) error { return nil }
-
-type panicHookModule struct {
-	stubCoreModule
-	panicOnEnter bool
-	panicOnExit  bool
-}
-
-func (p *panicHookModule) OnPhaseEnter(_ context.Context, _ Phase) error {
-	if p.panicOnEnter {
-		panic("enter hook boom")
-	}
-	return nil
-}
-
-func (p *panicHookModule) OnPhaseExit(_ context.Context, _ Phase) error {
-	if p.panicOnExit {
-		panic("exit hook boom")
-	}
-	return nil
-}
-
-func TestPhaseEngine_StartRollsBackWhenInitialEnterFails(t *testing.T) {
-	hook := &failingEnterHookModule{stubCoreModule: stubCoreModule{name: "hook"}, failPhase: "intro"}
-	pe, audit := newTestPhaseEngine(t, []Module{hook}, testPhaseDefinitions)
-
-	err := pe.Start(context.Background(), nil)
-	if err == nil {
-		t.Fatal("expected start enter failure")
-	}
-	if pe.started {
-		t.Fatal("engine should not remain started after initial enter failure")
-	}
-	if phase := pe.CurrentPhase(); phase != nil {
-		t.Fatalf("CurrentPhase after failed start = %#v, want nil", phase)
-	}
-	if got := len(audit.eventsOfType("engine.started")); got != 0 {
-		t.Fatalf("engine.started audits = %d, want 0", got)
-	}
-}
-
-func TestPhaseEngine_AdvanceRollsBackCurrentWhenTargetEnterFails(t *testing.T) {
-	hook := &failingEnterHookModule{stubCoreModule: stubCoreModule{name: "hook"}, failPhase: "invest"}
-	pe, audit := newTestPhaseEngine(t, []Module{hook}, testPhaseDefinitions)
-	ctx := context.Background()
-	if err := pe.Start(ctx, nil); err != nil {
-		t.Fatalf("Start: %v", err)
-	}
-	defer pe.Stop(ctx)
-
-	hasNext, err := pe.AdvancePhase(ctx)
-	if err == nil || hasNext {
-		t.Fatalf("AdvancePhase should fail without advancing, hasNext=%v err=%v", hasNext, err)
-	}
-	if got := pe.CurrentPhase().ID; got != "intro" {
-		t.Fatalf("CurrentPhase after failed advance = %s, want intro", got)
-	}
-	if got := pe.CurrentRound(); got != 1 {
-		t.Fatalf("CurrentRound after failed advance = %d, want 1", got)
-	}
-	if got := len(audit.eventsOfType("phase.advanced")); got != 0 {
-		t.Fatalf("phase.advanced audits = %d, want 0", got)
-	}
-}
-
-func TestPhaseEngine_SkipRollsBackCurrentWhenTargetEnterFails(t *testing.T) {
-	hook := &failingEnterHookModule{stubCoreModule: stubCoreModule{name: "hook"}, failPhase: "vote"}
-	pe, audit := newTestPhaseEngine(t, []Module{hook}, testPhaseDefinitions)
-	ctx := context.Background()
-	if err := pe.Start(ctx, nil); err != nil {
-		t.Fatalf("Start: %v", err)
-	}
-	defer pe.Stop(ctx)
-
-	err := pe.SkipToPhase(ctx, "vote")
-	if err == nil {
-		t.Fatal("expected SkipToPhase enter failure")
-	}
-	if got := pe.CurrentPhase().ID; got != "intro" {
-		t.Fatalf("CurrentPhase after failed skip = %s, want intro", got)
-	}
-	if got := len(audit.eventsOfType("phase.skipped")); got != 0 {
-		t.Fatalf("phase.skipped audits = %d, want 0", got)
-	}
-}
-
-func TestPhaseEngine_RequiredModuleDispatchUsesPanicIsolation(t *testing.T) {
-	pm := &panicModule{stubCoreModule: stubCoreModule{name: "information_delivery"}}
-	pe, audit := newTestPhaseEngine(t, []Module{pm}, testPhaseDefinitions)
-	ctx := context.Background()
-	if err := pe.Start(ctx, nil); err != nil {
-		t.Fatal(err)
-	}
-	defer pe.Stop(ctx)
-
-	err := pe.DispatchAction(ctx, PhaseActionPayload{Action: ActionDeliverInformation})
-	if err == nil {
-		t.Fatal("expected panic-isolated error")
-	}
-	if got := len(audit.eventsOfType("module.panic")); got != 1 {
-		t.Fatalf("module.panic audits = %d", got)
-	}
-}
-
-func TestPhaseEngine_PhaseHookPanicIsolation(t *testing.T) {
-	pm := &panicHookModule{stubCoreModule: stubCoreModule{name: "hook_panic"}, panicOnEnter: true}
-	pe, audit := newTestPhaseEngine(t, []Module{pm}, testPhaseDefinitions)
-	err := pe.Start(context.Background(), nil)
-	if err == nil {
-		t.Fatal("expected hook panic to become error")
-	}
-	if got := len(audit.eventsOfType("module.panic")); got != 1 {
-		t.Fatalf("module.panic audits = %d", got)
-	}
-}
-
-func TestPhaseEngine_SkipToPhaseRunsExitHooks(t *testing.T) {
-	hook := &hookRecordingModule{stubCoreModule: stubCoreModule{name: "hook"}}
-	pe, _ := newTestPhaseEngine(t, []Module{hook}, testPhaseDefinitions)
-	ctx := context.Background()
-	if err := pe.Start(ctx, nil); err != nil {
-		t.Fatal(err)
-	}
-	defer pe.Stop(ctx)
-
-	if err := pe.SkipToPhase(ctx, "vote"); err != nil {
-		t.Fatal(err)
-	}
-	if len(hook.exited) != 1 || hook.exited[0] != "intro" {
-		t.Fatalf("skip should exit old phase, exited=%#v", hook.exited)
-	}
-	if len(hook.entered) != 2 || hook.entered[1] != "vote" {
-		t.Fatalf("skip should enter target phase, entered=%#v", hook.entered)
-	}
-}
